refactor(cl): convert AudioESTestCase to mixin
